update code
the menu is visible and is added a button to go to cart where is possible to delete and add request to the order. 
the checkout button erase the cart and is sent to the menu. the add button is limited to no more than 9 and no less than 1 to choose a product.